+++ ft_memcmp.c
@@ -12,9 +12,9 @@
 
 #include <stdlib.h>
 
-int		ft_memcmp(const void *s1, const void *s2, int n)
+int		ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int		i;
+	size_t			i;
 	unsigned char	*str1;
 	unsigned char	*str2;
