+++ ft_memmove.c
@@ -11,13 +11,14 @@
 /* ************************************************************************** */
 
 #include <stdlib.h>
+#include <string.h>
 
-void	*ft_memmove(void *dest, const void *src, int n)
+void	*ft_memmove(void *dest, const void *src, size_t n)
 {
 	char	*de;
 	char	*sr;
 	char	*temp;
-	int		i;
+	size_t	i;
 	
 	de = (char*)dest;
 	sr = (char*)src;
